+++ packages/shared/components/modals/ProfileActions.svelte
@@ -2,8 +2,7 @@
     import { fade } from 'svelte/transition'
     import { Button, DeveloperIndicatorPill, HR, Icon, Modal, Text, Toggle } from 'shared/components'
     import { localize } from '@core/i18n'
-    import { LedgerConnectionState } from '@lib/typings/ledger'
-    import { ledgerDeviceStatus, getLedgerDeviceStatus } from 'shared/lib/ledger'
+    import { LedgerConnectionState, ledgerDeviceStatus, getLedgerDeviceStatus } from '@core/ledger'
     import { popupState, openPopup } from 'shared/lib/popup'
     import { openSettings } from '@core/router'
     import { diffDates, getBackupWarningColor, getInitials, isRecentDate } from 'shared/lib/helpers'
