+++ packages/shared/components/popups/ExportTransactionHistory.svelte
@@ -5,7 +5,7 @@
     } from 'shared/lib/transactionHistory'
     import { Button, PasswordInput, Spinner, Text } from 'shared/components'
     import { Platform } from 'shared/lib/platform'
-    import { displayNotificationForLedgerProfile, isLedgerConnected } from 'shared/lib/ledger'
+    import { displayNotificationForLedgerProfile, ledgerDeviceStatus } from '@core/ledger'
     import { showAppNotification } from 'shared/lib/notifications'
     import { closePopup } from 'shared/lib/popup'
     import { activeProfile, isLedgerProfile, isSoftwareProfile } from '@core/profile'
@@ -27,7 +27,7 @@
 
             if ($isSoftwareProfile && $isStrongholdLocked) {
                 await setStrongholdPassword(password)
-            } else if ($isLedgerProfile && !isLedgerConnected()) {
+            } else if ($isLedgerProfile && !$ledgerDeviceStatus.connected) {
                 isBusy = false
 
                 displayNotificationForLedgerProfile('warning')
