+++ packages/shared/components/popups/LedgerAddress.svelte
@@ -1,6 +1,5 @@
 <script lang="typescript">
     import { Animation, Text } from 'shared/components'
-    import { formatAddressForLedger } from 'shared/lib/ledger'
     import { showAppNotification } from 'shared/lib/notifications'
     import { closePopup, popupState } from 'shared/lib/popup'
     import { onMount } from 'svelte'
@@ -38,5 +37,5 @@
 </div>
 <div class="rounded-lg bg-gray-50 dark:bg-gray-800 p-5 text-center">
     <Text type="h5" highlighted classes="mb-2">{locale('general.receiveAddress')}</Text>
-    <Text type="pre">{formatAddressForLedger(address)}</Text>
+    <Text type="pre">{address}</Text>
 </div>
