+++ packages/shared/components/popups/ManageAccountPopup.svelte
@@ -4,7 +4,7 @@
     import { localize } from '@core/i18n'
     import { activeProfile, isLedgerProfile, isSoftwareProfile } from '@core/profile'
     import { selectedAccount, tryEditSelectedAccountMetadata, validateAccountName } from '@core/account'
-    import { promptUserToConnectLedger } from '@lib/ledger'
+    import { promptUserToConnectLedger } from '@core/ledger'
     import { closePopup, openPopup } from '@lib/popup'
 
     export let error = ''
