+++ packages/shared/components/popups/WalletFinderPopup.svelte
@@ -2,7 +2,7 @@
     import { Button, KeyValueBox, Spinner, Text, TextHint } from 'shared/components'
     import { closePopup, openPopup } from 'shared/lib/popup'
     import { showAppNotification } from 'shared/lib/notifications'
-    import { displayNotificationForLedgerProfile, isLedgerConnected } from 'shared/lib/ledger'
+    import { displayNotificationForLedgerProfile, ledgerDeviceStatus } from '@core/ledger'
     import { localize } from '@core/i18n'
     import {
         activeAccounts,
@@ -59,7 +59,7 @@
                 error = ''
                 isBusy = true
 
-                if ($isLedgerProfile && !isLedgerConnected()) {
+                if ($isLedgerProfile && !$ledgerDeviceStatus.connected) {
                     isBusy = false
 
                     displayNotificationForLedgerProfile('warning')
