+++ packages/shared/lib/core/account/actions/tryEditSelectedAccountMetadata.ts
@@ -1,6 +1,6 @@
 import { localize } from '@core/i18n'
 import { isLedgerProfile, updateActiveAccountMetadata } from '@core/profile'
-import { displayNotificationForLedgerProfile } from '@lib/ledger'
+import { displayNotificationForLedgerProfile } from '@lib/core/ledger'
 import { showAppNotification } from '@lib/notifications'
 import { get } from 'svelte/store'
 import { IAccountMetadata } from '../interfaces'
