+++ packages/shared/lib/core/ledger/actions/displayNotificationForLedgerProfile.ts
@@ -1,0 +1,43 @@
+import { get } from 'svelte/store'
+
+import { NotificationType } from '@lib/typings/notification'
+import { isNewNotification, showAppNotification } from '@lib/notifications'
+
+import { ledgerDeviceStatus } from '../stores'
+import { getLedgerDeviceStatus } from './getLedgerDeviceStatus'
+
+import { localize } from '@core/i18n'
+
+export function displayNotificationForLedgerProfile(
+    notificationType: NotificationType = 'error',
+    allowMultiple: boolean = true,
+    checkDeviceStatus: boolean = false,
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    error: any = null
+): string {
+    let notificationId
+
+    const _notify = () => {
+        const status = get(ledgerDeviceStatus)
+        const allowedToNotify = allowMultiple ? true : isNewNotification(notificationType)
+
+        const shouldNotify = allowedToNotify
+        if (shouldNotify) {
+            const stateErrorMessage = localize(`error.ledger.${status.connectionState}`)
+            const errorMessage = error?.error ? localize(error.error) : error
+            const message = error ? errorMessage : stateErrorMessage
+            notificationId = showAppNotification({
+                type: notificationType,
+                message,
+            })
+        }
+    }
+
+    if (checkDeviceStatus) {
+        getLedgerDeviceStatus()
+    } else {
+        _notify()
+    }
+
+    return notificationId
+}
