+++ packages/shared/lib/core/ledger/actions/openLedgerNotConnectedPopup.ts
@@ -1,0 +1,19 @@
+import { get } from 'svelte/store'
+import { openPopup, popupState } from '@lib/popup'
+
+export function openLedgerNotConnectedPopup(
+    cancel: () => void = () => {},
+    poll: () => void = () => {},
+    overridePopup: boolean = false
+): void {
+    if (!get(popupState).active || overridePopup) {
+        openPopup({
+            type: 'ledgerNotConnected',
+            hideClose: true,
+            props: {
+                handleClose: () => cancel(),
+                poll,
+            },
+        })
+    }
+}
