+++ packages/shared/lib/core/ledger/constants/ledger-status-poll-interval.constant.ts
@@ -1,0 +1,1 @@
+export const LEDGER_STATUS_POLL_INTERVAL = 1500
