+++ packages/shared/lib/core/ledger/constants/use-ledger-simulator.constant.ts
@@ -1,0 +1,1 @@
+export const USE_LEDGER_SIMULATOR = false
