+++ packages/shared/lib/core/ledger/enums/ledger-app-name.enum.ts
@@ -1,0 +1,4 @@
+export enum LedgerAppName {
+    BOLOS = 'BOLOS',
+    SHIMMER = 'Shimmer',
+}
