+++ packages/shared/lib/core/ledger/interfaces/ledger-connection-state.interface.ts
@@ -1,0 +1,8 @@
+export enum LedgerConnectionState {
+    AppNotOpen = 'appNotOpen',
+    Connected = 'connected',
+    Locked = 'locked',
+    NotDetected = 'notDetected',
+    OtherConnected = 'otherConnected',
+    MnemonicMismatch = 'mnemonicMismatch',
+}
