+++ packages/shared/lib/core/ledger/interfaces/ledger-extended-status.interface.ts
@@ -1,0 +1,6 @@
+import { LedgerStatus } from './ledger-status.interface'
+import { LedgerConnectionState } from './ledger-connection-state.interface'
+
+export interface LedgerExtendedStatus extends LedgerStatus {
+    connectionState: LedgerConnectionState
+}
