+++ packages/shared/lib/core/ledger/interfaces/ledger-status.interface.ts
@@ -1,0 +1,9 @@
+export interface LedgerStatus {
+    connected: boolean
+    locked: boolean
+    blindSigningEnabled: boolean
+    app?: {
+        name: string
+        version: string
+    }
+}
