+++ packages/shared/lib/core/ledger/stores/is-polling-ledger-device-status.ts
@@ -1,0 +1,3 @@
+import { writable } from 'svelte/store'
+
+export const isPollingLedgerDeviceStatus = writable<boolean>(false)
