+++ packages/shared/lib/core/ledger/stores/ledger-device-status.store.ts
@@ -1,0 +1,9 @@
+import { writable } from 'svelte/store'
+import { LedgerExtendedStatus, LedgerConnectionState } from '../interfaces'
+
+export const ledgerDeviceStatus = writable<LedgerExtendedStatus>({
+    connected: false,
+    locked: false,
+    blindSigningEnabled: false,
+    connectionState: LedgerConnectionState.NotDetected,
+})
