+++ packages/shared/lib/core/ledger/utils/isLedgerError.ts
@@ -1,0 +1,15 @@
+export function isLedgerError(error: { name; type }): boolean {
+    if (!error) return false
+
+    let errorType: string = ''
+    switch (typeof error) {
+        case 'object':
+            errorType = error.type || error.name
+            break
+        case 'string':
+            errorType = error
+            break
+    }
+
+    return errorType?.slice(0, 6) === 'Ledger'
+}
