+++ packages/shared/lib/core/profile/actions/active-profile/logout.ts
@@ -1,5 +1,5 @@
 import { resetRouters } from '@core/router'
-import { stopPollingLedgerStatus } from '@lib/ledger'
+import { stopPollingLedgerStatus } from '@lib/core/ledger'
 import { closePopup } from '@lib/popup'
 import { get } from 'svelte/store'
 import { destroyProfileManager } from '@core/profile-manager'
