+++ packages/shared/lib/typings/index.ts
@@ -2,11 +2,10 @@ import * as address from './address'
 import * as chart from './chart'
 import * as currency from './currency'
 import * as events from './events'
-import * as ledger from './ledger'
 import * as message from './message'
 import * as migration from './migration'
 import * as notification from './notification'
 import * as utils from './utils'
 import * as wallet from './wallet'
 
-export { address, chart, currency, events, ledger, message, migration, notification, utils, wallet }
+export { address, chart, currency, events, message, migration, notification, utils, wallet }
