+++ packages/shared/routes/onboarding/views/CongratulationsView.svelte
@@ -5,10 +5,9 @@
     import { mobile } from '@core/app'
     import { convertToFiat, currencies, exchangeRates, formatCurrency } from '@lib/currency'
     import { Platform } from '@lib/platform'
-    import { promptUserToConnectLedger } from '@lib/ledger'
     import { LOG_FILE_NAME, migration, migrationLog, resetMigrationState, totalMigratedBalance } from '@lib/migration'
     import { onboardingRouter, ledgerSetupRouter } from '@core/router'
-    import { LedgerAppName } from '@lib/typings/ledger'
+    import { LedgerAppName, promptUserToConnectLedger } from '@core/ledger'
     import { formatUnitBestMatch } from '@lib/units'
     import { AvailableExchangeRates, CurrencyTypes } from '@lib/typings/currency'
     import { localize } from '@core/i18n'
