+++ packages/shared/routes/onboarding/views/ledger-setup/views/AccountIndexView.svelte
@@ -2,11 +2,10 @@
     import { onDestroy, createEventDispatcher } from 'svelte'
     import { Animation, Button, Number, OnboardingLayout, Spinner, Text, Toggle, Icon } from 'shared/components'
     import { localize } from '@core/i18n'
-    import { displayNotificationForLedgerProfile, promptUserToConnectLedger } from '@lib/ledger'
     import { ADDRESS_SECURITY_LEVEL, hardwareIndexes } from '@lib/migration'
     import { Platform } from '@lib/platform'
     import { popupState } from '@lib/popup'
-    import { LedgerAppName } from '@lib/typings/ledger'
+    import { LedgerAppName, displayNotificationForLedgerProfile, promptUserToConnectLedger } from '@core/ledger'
 
     const dispatch = createEventDispatcher()
     const min = 0
