+++ packages/shared/routes/onboarding/views/ledger-setup/views/GenerateNewAddressView.svelte
@@ -2,7 +2,7 @@
     import { Animation, Button, Icon, OnboardingLayout, Spinner, Text } from 'shared/components'
     import { localize } from '@core/i18n'
     import { ledgerSetupRouter } from '@core/router'
-    import { formatAddressForLedger, promptUserToConnectLedger } from '@lib/ledger'
+    import { formatAddressForLedger, promptUserToConnectLedger } from '@core/ledger'
 
     let newAddress = ''
     let busy = false
