+++ packages/shared/routes/onboarding/views/ledger-setup/views/SwitchAppsView.svelte
@@ -2,8 +2,7 @@
     import { createEventDispatcher } from 'svelte'
     import { Animation, Button, Icon, OnboardingLayout, Text } from 'shared/components'
     import { localize } from '@core/i18n'
-    import { promptUserToConnectLedger, displayNotificationForLedgerProfile, ledgerDeviceState } from '@lib/ledger'
-    import { LedgerAppName, LedgerDeviceState } from '@lib/typings/ledger'
+    import { promptUserToConnectLedger, displayNotificationForLedgerProfile, LedgerAppName } from '@core/ledger'
 
     export let busy = false
 
@@ -19,8 +18,7 @@
         }
 
         function _onConnected(): void {
-            if ($ledgerDeviceState !== LedgerDeviceState.LegacyConnected) _onCancel()
-            else dispatch('next')
+            dispatch('next')
         }
 
         promptUserToConnectLedger(_onConnected, _onCancel)
